Actualizo los cambios de toneladas y hectareas

+++ SustitucionMOAModel/Models/DataAgro/NuevoAcopio.cs
@@ -8,7 +8,7 @@ namespace SustitucionMOAModel.Models.DataAgro
 {
     public class NuevoAcopio
     {
-        public int Toneladas { get; set; }
+        public float Toneladas { get; set; }
         public int LocalidadID { get; set; }
         public bool ArrendaPropia { get; set; }
         public int CampañaID { get; set; }
